~ changed autoaccount to use role enumerations instead of integer values ~ FileLogStream wasn't appending a new line at the end of the log

+++ Server/ClusterController/Database/AccountDB.cs
@@ -90,7 +90,7 @@ namespace ClusterControler.Database
             }
         }
 
-        public void CreateAccount(string accountName, string accountPassword, int role)
+        public void CreateAccount(string accountName, string accountPassword, ulong role)
         {
             accountName = Database.DoEscapeString(accountName);
